pyproject.toml
[tool.poetry]
name = "ohtu-miniprojekti-2021"
version = "0.1.0"
description = ""
authors = ["uradora <meri.rajamaki@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Jinja2 = "^2.11.3"
gunicorn = "^20.1.0"
selenium = "^3.141.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
coverage = "^5.5"
robotframework = "^4.0.1"
robotframework-seleniumlibrary = "^5.1.3"
pylint = "^2.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
